requirements.txt
numpy
pandas
matplotlib
plotly
pillow
streamlit
streamlit-ws-localstorage==1.0.0
websockets